requirements.txt
#  Python requirements for Door_Monitor.py
adafruit-io
paho-mqtt
RPi.GPIO

